requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
werkzeug==2.3.7
tinydb==4.7.1
torch==2.0.0
torchvision==0.15.0
pillow==8.4.0
numpy==1.23.5
websockets==12.0